+++ game_patch/main/main.cpp
@@ -209,6 +209,7 @@ FunHook<void(bool)> level_init_post_hook{
         reset_achievement_state_info();
         multi_level_init_post_gametypes();
         apply_geoable_flags();
+        apply_breakable_materials();
 
         if (!rf::is_dedicated_server) {
             explosion_flash_lights_level_init();